Fix various test errors because of invalid type check due to split HttpException
