+++ vertx-web-auth-common/src/main/java/io/vertx/ext/web/common/HttpException.java
@@ -28,11 +28,7 @@ import io.netty.handler.codec.http.HttpResponseStatus;
  *
  * @author <a href="mailto:plopes@redhat.com">Paulo Lopes</a>
  */
-public class HttpException extends RuntimeException {
-
-  public static final HttpException UNAUTHORIZED = new HttpException(401);
-  public static final HttpException BAD_REQUEST = new HttpException(400);
-  public static final HttpException BAD_METHOD = new HttpException(405);
+public abstract class HttpException extends RuntimeException {
 
   private final int statusCode;
   private final String payload;
